refactor: Make all functions async
This way none of them are run on main thread and as such if a function
panics, it will only crash that specific thread allowing the rest of the
application to survive.

The only exception is `force_panic()` as its purpose is hard crashing
the application.

From the frontend side of things, `invoke()` is already async, so no
frontend changes are needed.