feat: Allow custom TF|2 paths for dev builds
The point behind this is to allow for custom paths for dev builds that
can be hardcoded to allow pointing to some (fake) Titanfall2 install.
That way when working on a machine with no proper Titanfall2 install
or one where it's not being auto-detected, the path can be temporarly
hardcoded by the developer to avoid having to select it on each
recompile of the backend.